Refactor evidence helpers into SRP modules
